+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Move code to expire deleted tokens from Redis into separate event listener, closes #111.
 
 ## 0.0.36 - 2023-10-29
 ### Added